revert dtype

+++ diffusion/evaluation/generate_images.py
@@ -62,7 +62,7 @@ class ImageGenerator:
         if not hf_model:
             self.model = model
         else:
-            self.model = AutoPipelineForText2Image.from_pretrained(model, use_auth_token=True).to('cuda')
+            self.model = AutoPipelineForText2Image.from_pretrained(model, torch_dtype=torch.float16, use_auth_token=True).to('cuda')
         self.dataset = dataset
         self.load_path = load_path
         self.local_checkpoint_path = local_checkpoint_path
